pyproject.toml
[build-system]
requires = ["setuptools>=77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "mozilla-django-oidc-db"
version = "0.25.1"
description = "A database-backed configuration for mozilla-django-oidc"
authors = [
    {name = "Maykin Media", email = "support@maykinmedia.nl"}
]
readme = "README.rst"
license = "MIT"
license-files = ["LICENSE"]
keywords = ["OIDC", "django", "database", "authentication"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Framework :: Django :: 5.2",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.12"
dependencies = [
    "django>=4.2",
    "django-jsonform>=2.12",
    "glom",
    "mozilla-django-oidc>=3.0.0",
    "typing-extensions>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/maykinmedia/mozilla-django-oidc-db"
Documentation = "https://mozilla-django-oidc-db.readthedocs.io/en/latest/"
"Bug Tracker" = "https://github.com/maykinmedia/mozilla-django-oidc-db/issues"
"Source Code" = "https://github.com/maykinmedia/mozilla-django-oidc-db"
Changelog = "https://github.com/maykinmedia/mozilla-django-oidc-db/blob/master/CHANGELOG.rst"

[project.optional-dependencies]
setup-configuration = [
    "django-setup-configuration>=0.11.0",
]
tests = [
    "psycopg",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "pytest-mock",
    "pytest-recording",
    "requests-mock",
    "factory-boy",
    "pyquery",
    "tox",
    "ruff",
    "pyright",
    "django-stubs",
]
docs = [
    "sphinx",
    "sphinx-rtd-theme",
]
release = [
    "bump-my-version",
]

[tool.setuptools.packages.find]
include = ["mozilla_django_oidc_db*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "testapp.settings"
markers = [
    "oidcconfig: keyword arguments for the OIDC config",
    "auth_request: additional configuration for the auth_request fixture",
    "callback_request: additional configuration for the callback_request fixture",
    "mock_backend_claims: claims to be returned by the mock backend fixture",
]

[tool.bumpversion]
current_version = "0.25.1"
files = [
    {filename = "pyproject.toml"},
    {filename = "README.rst"},
    {filename = "docs/conf.py"},
]

[tool.coverage.run]
branch = true
source = [
    "mozilla_django_oidc_db"
]
omit = [
    # migrations run while django initializes the test db
    "*/migrations/*",
]

[tool.coverage.report]
exclude_also = [
    "if (typing\\.)?TYPE_CHECKING:",
    "@(typing\\.)?overload",
    "class .*\\(.*Protocol.*\\):",
    "@(abc\\.)?abstractmethod",
    "raise NotImplementedError",
    "\\.\\.\\.",
    "\\bpass$",
]
omit = [
    "*/migrations/*",
]

[tool.pyright]
include = [
    "mozilla_django_oidc_db",
    "testapp",
    "tests",
]
ignore = [
    # at the time of writing, the type definitions in factory-boy are still broken
    "mozilla_django_oidc_db/tests/factories.py",
    # don't know what's going on here, in particular because it's based on Pydantic I'd
    # expect this to have zero type checking issues. TODO for later
    "mozilla_django_oidc_db/setup_configuration/",
]
# pythonVersion = "3.10"
pythonPlatform = "Linux"

[tool.ruff.lint]
extend-select = [
    "UP",  # pyupgrade
    "DJ",  # django
    "LOG", # logging
    "G",
    "I",   # isort
    "E",   # pycodestyle
    "F",   # pyflakes
    "PERF",# perflint
    "B",   # flake8-bugbear
]
extend-ignore = [
    "E501",  # a lot of line-length too long violations...
]

[tool.ruff.lint.isort]
combine-as-imports = true
section-order = [
    "future",
    "standard-library",
    "django",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"django" = ["django"]
